Update integration scheme in iflxcalc() and set layer boundary to 112.5m
Integration intervals for fluxes were updated to have interval centered around gridpoints, which is better compared to the previous case in which intervals were between gridpoints. Also changed boundary between layers A and B to be located between gridpoints at 110m and 115m, which reduces ambiguitiy in picking where a gridpoint belongs to one layer or another.

Verified that graphical output is equivalent to previous output (as it should be), and that integral calcs were simlar to those I showed during EXPORTS webinar 20200608.