Make calls to acdrange() with correct depth intervals
acdrange() calculates the autocorrelation function for a subset of the total POC data and does a linear regression on that function. Calls to the acdrange() should be specific to a layer. The correction ensures that no total POC data point is considered in multiple layers. Verified that graphical output is not drastically changed, neither are integrated fluxes or timescales in comparison to those presented during EXPORTS webinar 20200608.