Change prior for NPP in ML, Add tracer inventory calcs
The old prior for NPP in the ML (Gh in the model) was too high, according to a ccloser look at JFox's 14C NPP data. added inventory() to calculate tracer inventories for comparison with integrated fluxes and timescale calculations. Renamed flxep() to symfunceval() since it evaluates and symbolic function (with error propagation if desired), not just for calculation of fluxes.