Update sign convention in model equations
In both the ATI loop's calculation of model equations (f) and in the function that calculates residuals of model equations modresi(), sign convention is now more intuitive – sources are positive and sink terms are negative. Note sinking flux divergence has a negative sign in front of it. Verified that graphical output is unchanged, as it should be. Exception is plots of model residuals, which are now flipped about the y-axis.