Add calculation of integrated residuals to inventory()
The sum of all integrated fluxes (i.e. budget) for a given tracer in a given depth range should be equal to the integrated residuals for that depth range. inventory() can now calculate integrated residuals for the depth range provided as an argument. I also added a test at the end that shows these integrated residuals are equal to the budget "residuals" (defecits/surpluses).