Set ATI to converge following Murnane 1994
The convergence criterion described in Murnane 1994 allows the ATI to converge when the absolute change in each state variable between xk and xk+1 within an iteration is at most 1%. This criterion is simpler and allows the code to be more consistent with previous studies. I verified that graphical output and budgets are sufficiently similar to results from inversions that used the previous convergence criterion.