Update convergence criteria
Before, I was checking for the max of abs(xkp1-xk)/xk and comparing to a threshold (0.01 as in Murnane 94). However, this would not always return the desired result if the maximum change is observed when a value in xk is negative. Instead, I changed it to check abs((xkp1-xk)/xk), which ensures that the expression is always positive. I also compared this to a method comparing abs(xkp1-x1) to 0.01*abs(xk) for every corresponding values of xk and xkp1, and comfirmed that they are indeed equivalent.