Remove normalization of state vector by prior values
I also changed the pdelt variable from 0.01 to 0.0001. When I kept it the same, the graphical output and budget residuals were slightly different compared to the case with normalization. Convergence in the unnormalized case happened in only 9 steps, compared to ~50 in the normalized case. This is becuase maintaining the same pdelt imposes a relatively stricter constraint on the convergence criteria when the state vector is normalized. When I changed pdelt to 0.0001 for the unnormalized case, graphical output compared to the normalized pdelt = 0.01 case was identical by inspection. Budget residuals were almost identical (exception is Pl residuals, identical to 3 decimal places). Convergence happened in ~70 steps.