Add gentargets.py, targets.csv, twnxP_ts.py
The goal for this commit was to generate a number (currently 10) of groups of target values with which to generate pseudo-data and perform twin experiments. gentargets.py generates these target values and outputs them in a csv file (targets.csv) with three columns: group (an index for the set or group of target values that a certain value belongs to), param, and val(ue). Target values for a given param are randomly selected from a lognormal distribution with mean equal to the prior estimate and stdev equal to prior error. Becuase of the lognormal distribution, this tends to produce target values that are less than the prior estimate but also sometimes produces values that are much greater due to the long tail of the distribution. Perhaps a better approach would be to pick positive values uniformly distributed about one standard deviation of the prior estimate, as our prior errors seem pretty generous.

twnxP_ts.py then performs twin experiments after reading in the target values from targets.csv. It currently uses a noiseless estimate of total POC as the sum of Ps and Pl from the pseudo-data, since I discovered that issues with convergence arise if the total POC data is too noisy (it is not very noisy in the real data, after all). However, there are commented-out lines showing how I was previously selecting noisy Pt values. Figures are automatically saved out, with the suffix indicated the group of target values that it belongs to (e.g. _g2.png belongs to group 2).