Perform randomized twin experiments with target values drawn from uniform distributions
gentargets.py was modified to pick target values for rate parameters that are uniformly distributed about one standard deviation above and below the prior estimates. In the case where the standard deviation is greater than the mean etstimate, 0 is used as the lower bound for the uniform distribution. Because gentargets.py was changed, a new targets.csv was created, but I saved the old one as targets_oldlognormal.csv. I also removed a bunch of whitespace from twnxP_ts.py.