Update nested hierarchy of pdi variable in invP_v17.py and twnxP.py
pdi is a dictionary that stores information for rate parameters. I restructured it to be cleaner due to the fact that each rate parameter only has one prior estimate, for example, and before, there was a prior estimate for each independent layer within a given parameter key. This change reduces redundancy in dictionary branches. Verified that output for invP_v17.py was unchanged; that for twnxP.py changed since pseudo-data for total POC are still have randomly generated noise, which results in different rate parameter estimates every run. Given this difference, the change in output was minor enough to not warrant concern.