Move twnxP_ts.py to invPOC_misc (local directory)
Results indicate that choosing randomly selected sets of target values more likely than not chooses sets that produce pseudo-data that are not representative of real data, and/or do not allow the ATI to converge. We decided to simply use a twin experiment with targets set to posterior estimates from the real inversion for model validation.