Calculate residuals for total POC mass balance equations
Residuals for Pt equations (calculated as difference between cp-derived Pt and sum of inverted Ps and Pl estimates) are now calculated. Depth profiles for Pt residuals are included as a third panel in the residprofs plots. Residuals are also included in the second panel of the PDFs and CDFs, which shows residuals for all model equations.