Create pyrite.py and input data file, pyrite_data.xlsx
This is the beginning of an object-oriented approach to invP.py, which I'm rebranding as the Pyrite (Particle cYcling Rates from Inversion of Tracers in the ocEan) model. In this commit, I create a class file and define the __init__ method as well as some other basic class methods (load_data, unpack_tracers, define_params, process_npp_data). I verified that the prior estimates that come out of process_npp_data match those from invP.py.