Add indexes of variable names in the model, PyriteModelRun object to run model
The variable indexes are similar to those used in invP.py, and are necessary to keep track of where state elements are stored in the state vector, for example. Up to this point, PyriteModelRun has methods to define the model error equation matrix and run the ATI given a gamma value. Includes calculation of Jacobian (F) and vector of model equations (f). For gamma = 0.02, I verified that the evolution of the cost function and convergence criteria are the same as in invP.py, including that both convereged in the same number of steps. Next step is to extract posterior estimates of state elements, calculate their errors, and plot the posterior estiamtes to verify that they are the same as those produced by invP.py.