Allow depth-varying model parameters to only vary between 0-100m and 100-500m
There are still 7 layers, and one POCS(L) equation for each layer. There are just fewer rate parameters compared to the previous commit, as the depth-varying ones are only allowed to vary between the two depth horizons mentioned above. Model converges for gamma = 0.02, but not 0.01. This commit includes a plot that compares rate parameters estimated from this inversion to those from the previous model with objective interpolation.