Rewrite extract_equation_variables() and equation_builder()
zone_name was being reassigned in extract_equation_variables(), which changed the zone_name for all subsequent variables in the equation being considered. This is no good. Also, ws and ws-, e.g., were pointing to the same state element (ws_LEZ) in some equations, such as that for the layer spanning from 30-50m. They were being treated as seperate free symbols, which was problematic in calculating the Jacobian. In this edit, I assign the zone explicitly in the symbolic variable names in equation_builder, which makes the logic in extract_equation_variables() much more simple and less error-prone. Now, there is consistency in the Jacobian calculation between different runs at each iteration step, which was not the case before. Param and POC results are plotted, and param results are compared to that from the inversion from the first paper.