Add DVM via a B3 parameter
B3 removes Ps from zone A (mixed layer) and injects it into Pl in zone F (200 - 330m). In the Pl equation it is multiplied by "alpha" which represents the fraction of C consumed that is actually egested. Currently, alpha = 0.1. The prior estimate for B3 comes from Murnane's B2 from OSP, which is 0.8 +- 0.9 y^-1. Converges for gamma_POC = 0.01 and not for 0.005 as before adding DVM. Note that the code as is has a fixed gamma value for Ti (2.5), and is not actually designed to have an inversion with Ti data. I just kept the Ti-relevant code in there because it was easier than taking it out, and keeping it should have no outcome on the result of the POC-only inversion.