Add option to include or exclude dvm as keyword arg in PyriteModel
Keeping alpha = 0.1, see little difference in POC profile and rate parameter estimates when we include or exclude dvm. Current prior for B3 is still from Murnane, which is about 0.0022 d^-1 (0.8 += 0.9 y^-1). The model still converges with and without dvm when gamma = 0.01. With dvm, the estimate of B3 is not very different from the prior, and has larger error bars. I tried increasing the B3 prior to 0.01 d^-1 with 100% error to see if it would more drastically change the estimates when dvm is included, but seems like it doesn't. A B3 prior = 0.1 (with 100% error) or higher did not produce a converging solution. I suspect that the flux represented by this term is probably too small to make a difference in the results given the current assumptions. Next step will be to calculate the fluxes to reinforce this suspicion.