Calculate integrated fluxes and write results to output file
Fluxes are integrated in each of the 7 layers (A-G), but also in the LEZ (A-C) and UMZ (D-G). When the original prior for B3 is used (0.8 += 0.9 y^-1 ~ 0.0022 d^-1), the integrated fluxes between the dvm and no dvm case look pretty similar, and the integratred dvm fluxes are small relative to the other fluxes. In addition, the posterior POC profiles and parameter profiles look similar. When the prior is increased to 0.03 d^-1 with 100% error, the dvm fluxes increase but posterior POC profiles and rate parameter estimates still look similar.