Add DVM
Ps is consumed in the LEZ, and excreted in the UMZ as Pl as a function of depth according to a sin function (see notes from Olivier). Current prior for B3 is 0.05 d^-1 with 100% error, which results in a converging solution if gamma = 0.01. Raising prior to 0.1 d^-1 with 100% error does not allow model to converge.