Hard constraint with no logs, use priors from NABE or OSP
The error in model equations is represented explicitly as a term in each of the model equations, such that all equations are satisfied perfectly without need for a Cf matrix. No lognormal distributions are used. Has an an option to use priors for B2p and Bm2 based on either NABE or OSP estimates from Murnane. With both prior sets, model converges when gamma = 0.5. Gamma modulates the magnitude of error in the Ps and Pl equations. Error is proportional (via gamma) to integrated NPP in the mixed layer.