Add check_equation_residuals()
When the estimate of x is plugged back into the equations, we should get 0 for the sum of all terms. This is indeed the case for the most part. However, I get numbers on the order of 10^-6 to 10^-4 in some of the POCL equations (but not for EZ with NABE priors, those range from 10^-9 to 10^-12). Note here I'm talking about the inversion where gamma and relative error both = 0.5. The only guess that I have for this is rounding error, especially because the POCL equations (especially those in the UMZ) have many terms with a wide range of magnitudes due to the DVM flux.