Remove cp data
These data do not improve our results in the layer model enough to justify their inclusion. Also, the RMSE of the semilog fit of PT and cp data was small compared to some of the residuals of the fit (at high cp values). Becasue the RMSE was used to quantify the error in the residuals of the PT equations, this likely resulted in underestimation of such residuals. I also deleted some unused imports.