Calculate Lp and Po priors with errors propagated formally
I think the errors for Po (surface production) are overestimated, as they are propagated from NPP and Lp estimates, which have high variability (i.e. error)  throughout the transect. This is especially due to high production at the first few stations. It might be better to start with a relative error, say 25%, for the Lp and Po priors.