Super station inversions with upper 600m data
Relative error of 100% for all rate parameters except for Pdot and Lp. The same prior error for residuals is applied to all stations, which is equal to a gamma value of 0.1 multiplied by the station averages of the products of Pdot prior and MLD at each station (at which these two quantities are available across the transect, not just the super stations). All inversions converge, except for SP inversions at stations 8 and 29. POC profiles are well satisfied. Not much adjustment or improvement in precision for the uniform rate paramters. Residuals are generally positive and decrease towards 0 for PL, and generally negative and increase towards 0 for PS. For a given station, depth variation in residuals is consistent between NA and SP inversions, but errors in residuals for SP inversions are generally larger (probably to larger error in Bm2 and B2p priors). Stations that exhibit most depth variation in depth-varying parameters are stations 8 and 14. In stations 14, 23, and 35, there is some variation in Bm2 that does not seem to be consistent with features in any other profiles (POC or parameters). See figures for more.