Update calculation of residual prior errors (gp15)
For each station, EZ depths are estimated from Kd. Water column integrated NPP from satellite data is multiplied by the ratio of (i) the integral of the volumetric exponential NPP function from the surface to the base of the first layer, to (ii) the integral from the surface to the base of the EZ. The prior error for the residuals is set equal to the average of these products across all stations, which is then multiplied by gamma. Po priors for each station are also recalculated as the satellite NPP times the aforementioned ratio, divided by the thickness of the first layer. PPZ and MLD data are no longer needed.